multipaste: Clip contents when delete button is visible
This CL causes a clipboard history item's contents to be clipped when
the menu item's delete button is visible. The contents are clipped with
a cutout so that they do not overlap with the button.

Bitmap item: https://screenshot.googleplex.com/5VrUuNHsUmtZdRH
Text item: https://screenshot.googleplex.com/B4pu6urVqaAR2Xa

Note: In order to make it easier to see tell what coverage this CL adds,
the test changes introduced here account for the clipboard history
refresh being enabled or disabled but do not actually parameterize these
tests. The tests are moved to a parameterized suite in a follow-up CL.

Bug: b/267693870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4688865
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171429}

+++ ash/clipboard/views/clipboard_history_view_constants.h
@@ -42,10 +42,10 @@ constexpr int kLabelPreferredHeight = 20;
 constexpr int kImageViewPreferredHeight = 72;
 
 // The radius of the image view's rounded corners when offset by a background.
-constexpr int kImageBackgroundCornerRadius = 12;
+constexpr float kImageBackgroundCornerRadius = 12.f;
 
 // The radius of the image view's rounded corners when surrounded by a border.
-constexpr int kImageBorderCornerRadius = 4;
+constexpr float kImageBorderCornerRadius = 4.f;
 
 // The preferred size for an item's icon.
 constexpr gfx::Size kIconSize(20, 20);
